Update .zshrc to overwrite TERM only if it's not set
And here we go: first problem with plain overwriting of $TERM variable:
when starting 'tmux', it should set $TERM to 'screen-256color' and it
does, but after it runs zsh and $TERM is overwritten by .zshrc.

So from now $TERM will be set to 'xterm-256color' only if it doesn't
contain substring '256color' in it already.